requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alpaca-py==0.28.0
alpaca-trade-api==3.2.0
annotated-types==0.7.0
APScheduler==3.10.4
attrs==23.2.0
bayesian-optimization==1.5.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
deprecation==2.1.0
fonttools==4.53.1
frozenlist==1.4.1
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.1
msgpack==1.0.3
multidict==6.0.5
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sseclient-py==1.8.0
TA-Lib==0.4.32
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.19
websocket-client==1.8.0
websockets==10.4
yarl==1.9.4
